requirements.txt
numpy~=2.2.2
fastapi~=0.115.7
pydantic~=2.10.6
pillow~=11.1.0
uvicorn==0.22.0
python-multipart~=0.0.5